Split the test tree into cpp and mana directories
test/ held 47 mana scripts, two C++ test programs, their project files,
a Makefile and test.py in one flat directory, so the scripts the language
suite feeds to mana were indistinguishable from the fixture the program
image reader consumes.

Split by what each file is for rather than by extension. test/cpp/ holds
the two C++ suites, their vcxproj and Makefile, and ProgramImageTest.mn,
which exists only to be compiled into the image ProgramImageTest reads.
test/mana/ holds the 46 language cases and the test.py that drives them,
so that directory is exactly the suite it runs.

Relative paths follow the move: the C++ sources and project files reach
one level further up for compiler/ and runner/, mana.sln points at
test/cpp/, and the driver's test target and the Windows CI job address
each suite in its own directory. The gitignore rules for generated .h and
.md now cover the whole test tree instead of its top level only.
