+++ test/cpp/EmbeddingTest.vcxproj
@@ -19,7 +19,7 @@
     </ProjectConfiguration>
   </ItemGroup>
   <PropertyGroup Label="Globals">
-    <ProjectGuid>{9B4D6C20-18AF-4E75-8D31-2A6F0C7B41E5}</ProjectGuid>
+    <ProjectGuid>{3F2C8A17-5E64-4B92-A0D1-7C8B5E934F60}</ProjectGuid>
     <RootNamespace>mana</RootNamespace>
     <Keyword>Win32Proj</Keyword>
     <WindowsTargetPlatformVersion>10.0</WindowsTargetPlatformVersion>
@@ -180,29 +180,38 @@
     </Link>
   </ItemDefinitionGroup>
   <ItemGroup>
-    <ClInclude Include="..\runner\Actor.h" />
-    <ClInclude Include="..\runner\Buffer.h" />
-    <ClInclude Include="..\runner\common\Assert.h" />
-    <ClInclude Include="..\runner\common\FileFormat.h" />
-    <ClInclude Include="..\runner\common\Fault.h" />
-    <ClInclude Include="..\runner\common\Math.h" />
-    <ClInclude Include="..\runner\common\Memory.h" />
-    <ClInclude Include="..\runner\common\Noncopyable.h" />
-    <ClInclude Include="..\runner\common\OutputStream.h" />
-    <ClInclude Include="..\runner\common\Platform.h" />
-    <ClInclude Include="..\runner\common\Setup.h" />
-    <ClInclude Include="..\runner\common\Stopwatch.h" />
-    <ClInclude Include="..\runner\common\String.h" />
-    <ClInclude Include="..\runner\common\Type.h" />
-    <ClInclude Include="..\runner\Event.h" />
-    <ClInclude Include="..\runner\Function.h" />
-    <ClInclude Include="..\runner\Mana.h" />
-    <ClInclude Include="..\runner\Plugin.h" />
-    <ClInclude Include="..\runner\Stack.h" />
-    <ClInclude Include="..\runner\VM.h" />
+    <ClInclude Include="..\..\runner\Actor.h" />
+    <ClInclude Include="..\..\runner\Buffer.h" />
+    <ClInclude Include="..\..\runner\common\Assert.h" />
+    <ClInclude Include="..\..\runner\common\FileFormat.h" />
+    <ClInclude Include="..\..\runner\common\Fault.h" />
+    <ClInclude Include="..\..\runner\common\Math.h" />
+    <ClInclude Include="..\..\runner\common\Memory.h" />
+    <ClInclude Include="..\..\runner\common\Noncopyable.h" />
+    <ClInclude Include="..\..\runner\common\OutputStream.h" />
+    <ClInclude Include="..\..\runner\common\Platform.h" />
+    <ClInclude Include="..\..\runner\common\Setup.h" />
+    <ClInclude Include="..\..\runner\common\Stopwatch.h" />
+    <ClInclude Include="..\..\runner\common\String.h" />
+    <ClInclude Include="..\..\runner\common\Type.h" />
+    <ClInclude Include="..\..\runner\Event.h" />
+    <ClInclude Include="..\..\runner\Function.h" />
+    <ClInclude Include="..\..\runner\Mana.h" />
+    <ClInclude Include="..\..\runner\Plugin.h" />
+    <ClInclude Include="..\..\runner\Stack.h" />
+    <ClInclude Include="..\..\runner\VM.h" />
+    <ClInclude Include="..\..\compiler\Compiler.h" />
+    <ClInclude Include="..\..\compiler\ErrorHandler.h" />
+    <ClInclude Include="..\..\compiler\Path.h" />
+    <ClInclude Include="..\..\compiler\Version.h" />
   </ItemGroup>
   <ItemGroup>
-    <ClCompile Include="ProgramImageTest.cpp" />
+    <ClCompile Include="EmbeddingTest.cpp" />
+  </ItemGroup>
+  <ItemGroup>
+    <ProjectReference Include="..\..\compiler\manac.vcxproj">
+      <Project>{7c1a9f84-2b65-4e0c-9d3a-1f6e5b0c4a21}</Project>
+    </ProjectReference>
   </ItemGroup>
   <Import Project="$(VCTargetsPath)\Microsoft.Cpp.targets" />
   <ImportGroup Label="ExtensionTargets">
