+++ test/mana/TestBool01.mn
