+++ test/mana/TestConstDeclInvalidAssign.mn
