+++ test/mana/TestConstDeclInvalidCall.mn
