+++ test/mana/TestConstDeclInvalidDefine.mn
