+++ test/mana/TestConstDeclInvalidType.mn
