+++ test/mana/TestConstDeclOk.mn
