+++ test/mana/TestConstant01.mn
