+++ test/mana/TestExtendNamespaceResolution.mn
