+++ test/mana/TestForwardUsingNamespace.mn
