+++ test/mana/TestGlobalInit.mn
