+++ test/mana/TestNamespaceFqn.mn
