+++ test/mana/TestNativeFunctionBody.mn
