+++ test/mana/TestStatementSwitch.mn
