+++ test/mana/TestStruct01.mn
