+++ test/mana/TestStructMethod01.mn
