+++ test/mana/TestStructMethodNonStruct.mn
