+++ test/mana/TestStructMethodUnknown.mn
