+++ test/mana/TestStructNamespace.mn
