+++ test/mana/TestStructNativeMethod.mn
