+++ test/mana/TestUsingAmbiguousActor.mn
