+++ test/mana/TestUsingAmbiguousNamespaceSymbol.mn
