+++ test/mana/TestUsingAmbiguousType.mn
