+++ test/mana/TestUsingNamespace.mn
