+++ test/mana/TestUsingResolveNamespacePath.mn
