+++ test/mana/TestUsingResolveSymbolImport.mn
