+++ test/mana/TestUsingScopeOrder.mn
