+++ test/mana/TestUsingSymbol.mn
