+++ test/mana/TestVariable01.mn
