+++ test/mana/TestVariable02.mn
